moves mesoscope splitting to modules, removes a lot of excess stuff
