+++ src/ophys_etl/modules/mesoscope_splitting/metadata.py
