+++ src/ophys_etl/modules/mesoscope_splitting/schemas.py
