+++ src/ophys_etl/modules/mesoscope_splitting/tiff.py
