--- src/ophys_etl/pipelines/brain_observatory/README.md
@@ -1,10 +1,0 @@
-brain_observatory_utils
-===============================
-
-https://github.com/AllenInstitute/brain_observatory_utils
-
-Utitilies for dbrain observatory pipelines.
-
-Level of Support
-===============================
-We are not currently supporting this code, but simply releasing it to the community AS IS but are not able to provide any guarantees of support. The community is welcome to submit issues, but you should not expect an active response.
