--- src/ophys_etl/pipelines/brain_observatory/schemas/__init__.py
@@ -1,1 +1,0 @@
-from .ophys_platform_v2 import DeepscopeSchema, MesoscopeSchema
