--- src/ophys_etl/pipelines/brain_observatory/schemas/base.py
@@ -1,10 +1,0 @@
-from argschema.schemas import DefaultSchema
-from argschema.utils import load
-
-
-class BaseSchema(DefaultSchema):
-    _MANY = False
-
-    @classmethod
-    def load_validated(cls, data):
-        return load(cls(many=cls._MANY), data)
