--- src/ophys_etl/transforms/mesoscope_2p/__init__.py
@@ -1,1 +1,0 @@
-from ophys_etl.transforms.mesoscope_2p.tiff import MesoscopeTiff
